Major Changes: attn maps, reproducibility, configs
Attention Maps:
Under src/analysis we've added the code to produce the attention maps
that are within the latest release.

Reproducibility:
We've added more tracking for the arguments that include better tracking
of seeds. This is an attempt to make works as reproducible as possible
and doesn't come with many costs.

Configs:
We've changed the configs a little, mostly ensuring that there are
defaults and reducing the complexity of the main config. There are also
new options. The meta config file provides a description of all possible
options. A README was also added to explain some basic arguments
formatting.